Changes to be committed: 	modified:   actions/move.go 	modified:   pkg/help.go 	modified:   pkg/rebalance.go 	modified:   pkg/setting.go 	new file:   pkg/structs.go Move all struct tu struct.go. Small fix Cleared code
